refactor: LineMaker 삭제
