+++ src/main/java/ladder/domain/ladder/Ladder.java
@@ -18,7 +18,7 @@ public class Ladder {
         List<Line> lines = new ArrayList<>();
 
         for (int idx = 0; idx < height; idx++) {
-            List<Bar> bars = LineMaker.generate(count, booleanGenerator);
+            List<Bar> bars = generate(count, booleanGenerator);
             lines.add(new Line(bars));
         }
 
@@ -35,6 +35,27 @@ public class Ladder {
             currPosition = line.nextPosition(currPosition);
         }
         return currPosition;
+    }
+
+
+    public static List<Bar> generate(int playerCount, BooleanGenerator booleanGenerator) {
+        int lineSize = playerCount - 1;
+        List<Bar> line = new ArrayList<>();
+        boolean beforeValue = false;
+        for (int idx = 0; idx < lineSize; idx++) {
+            Bar currentBar = createBar(beforeValue, booleanGenerator);
+            line.add(currentBar);
+            beforeValue = currentBar.getValue();
+        }
+        return line;
+    }
+
+    private static Bar createBar(boolean beforeValue, BooleanGenerator booleanGenerator) {
+        if (beforeValue) {
+            return Bar.UNMOVABLE_BAR;
+        }
+
+        return Bar.getBar(booleanGenerator.generateBoolean());
     }
 
 }
