+++ src/test/java/ladder/domain/ladder/LadderTest.java
@@ -2,10 +2,8 @@ package ladder.domain.ladder;
 
 import ladder.domain.generator.BooleanGenerator;
 import ladder.domain.generator.MockBooleanGenerator;
-import org.junit.jupiter.api.Assertions;
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
+import ladder.domain.generator.RandomBooleanGenerator;
+import org.junit.jupiter.api.*;
 
 import java.util.List;
 
@@ -26,6 +24,28 @@ public class LadderTest {
     @DisplayName("사다리 생성을 테스트")
     void ladderInitiatorTest() {
         Assertions.assertDoesNotThrow(() -> Ladder.create(playerCount, heightOfLadder, valueGenerator));
+    }
+
+    @Test
+    @DisplayName("플레이어 수 -1 크기의 사다리가 생성되는지 확인한다.")
+    void generateLineSizeTest() {
+        assertThat(Ladder.create(playerCount, heightOfLadder, new RandomBooleanGenerator()).getLadder().get(0).getLine().size())
+                .isEqualTo(playerCount - 1);
+    }
+
+    @RepeatedTest(100)
+    @DisplayName("생성된 Line의 가로가 겹치지 않는지 확인한다.")
+    void generateLineTest() {
+        List<Bar> bars = Ladder.create(100, 1,new RandomBooleanGenerator())
+                .getLadder().get(0).getLine();
+
+        for (int idx = 0; idx < bars.size() - 1; idx++) {
+            Bar currentBar = bars.get(idx);
+            Bar nextBar = bars.get(idx + 1);
+
+            assertThat(!currentBar.getValue() || !nextBar.getValue())
+                    .isTrue();
+        }
     }
 
     @Test
